monitoring: split web.py
